VST3: Prevent a compiler warning

+++ modules/juce_audio_plugin_client/VST3/juce_VST3ModuleInfo.h
@@ -99,7 +99,7 @@ public:
                 << "\"New\": \"" << VST3::UID { newId }.toString() << "\", "
                 << "\"Old\": [";
 
-            for (int i = 0; i < oldIds.size(); ++i)
+            for (size_t i = 0; i < oldIds.size(); ++i)
             {
                 str << "\""
                     << std::hex
